lock keyword: let arg_{read,write} caller determine locking
Instead of unconditionally opening files with `lock=false`, this allows
the caller of `arg_read` or `arg_write` to pass a boolean `lock` keyword
argument to determine whether `IOStream` objects should do per-operation
locking or not. For `IOStream` objects created by opening a file, this
calls `open` with the `lock` keyword. For `IOStream` objects passed into
the `arg_read` or `arg_write` call, if `lock=false` the `IOStream` is
locked once for the whole operation and then the `_dolock` field is set
to `false` and restored upon termination of the block.